refactor: refactor AddCardPage component folder
