+++ src/pages/addCard/components/addCard/CardNumberInput.tsx
@@ -1,4 +1,4 @@
-import { Input, InputProps } from "../../../components/ui-kit/Input";
+import { Input, InputProps } from "../../../../components/ui-kit/Input";
 
 interface CardNumberInputProps {
   value: string;
