+++ src/pages/addCard/components/addCard/Dot.tsx
